fstree: move writer logic into a file of its own
Allow overriding it. No functional changes.
